requirements.txt
click
pyyaml
Jinja2
requests
